% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{accuracy_by_group}
\alias{accuracy_by_group}
\title{Accuracy by group}
\usage{
accuracy_by_group(df)
}
\arguments{
\item{df}{grouped data frame}
}
\value{
plot of education vs. pct. correct
}
\description{
Computes the percentage correct and average relative error for a grouped data set.
}

