% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{agg_plot_users}
\alias{agg_plot_users}
\title{Plot education vs accuracy}
\usage{
agg_plot_users(df)
}
\arguments{
\item{df}{grouped data frame}
}
\value{
plot of education vs. pct. correct
}
\description{
computes the overall accuracy score (pct. correct) vs. education level
}

