% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric Mean}
\usage{
geometric_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of integers / doubles}
}
\value{
The geometric mean of the vector.
}
\description{
Computes the geometric mean of a vector
}

