% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{max_sample_size}
\alias{max_sample_size}
\title{Maximum sample size}
\usage{
max_sample_size(id)
}
\arguments{
\item{id}{integer (domain id)}
}
\value{
the min nr of responses of any task in the domain
}
\description{
Computes the maximum sample size for a domain (min nr of responses of any task)
}

