% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{rank}
\alias{rank}
\title{Rank}
\usage{
rank(v, x, TrueValue, na.rm = FALSE)
}
\arguments{
\item{v}{vector of individual answers}

\item{x}{crowd estimate}

\item{TrueValue}{the correct answer}
}
\value{
The percentage of entries in v that are further from TrueValue than x.
}
\description{
Computes the percent of entries in v that x is better than
in terms of their distance from TrueValue.
}

