% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector of integers / doubles}
}
\value{
a vector with outlier observations removed
}
\description{
Removes outliers outside 95% confidence interval.
}

