% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{sample_size}
\alias{sample_size}
\title{Plot Accuracy vs Sample Size (MC)}
\usage{
sample_size(id, nr_simulations)
}
\arguments{
\item{id}{integer (domain id)}

\item{nr_simulations}{the number of random samples to average over}
}
\value{
plot of accuracy vs sample size
}
\description{
Computes the geometric mean of a vector for MC questions
}

