% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{sample_size_cont}
\alias{sample_size_cont}
\title{Plot Accuracy vs Sample Size (point estimate)}
\usage{
sample_size_cont(id, nr_simulations)
}
\arguments{
\item{id}{integer (domain id)}

\item{nr_simulations}{the number of samples to generate}
}
\value{
plot of accuracy vs sample size
}
\description{
Computes the geometric mean of a vector for point estimate questions
}

