% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing spaces}
\usage{
trim(x)
}
\arguments{
\item{x}{a vector of character}
}
\value{
the original string, without any trailing / leading spaces
}
\description{
Trim leading and trailing spaces
}

