% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/power.R
\name{truncated_geometric_mean}
\alias{truncated_geometric_mean}
\title{Truncated Geometric Mean}
\usage{
truncated_geometric_mean(x, trunc = 0.05, na.rm = FALSE)
}
\arguments{
\item{x}{a vector of integers / doubles}

\item{trunc}{real-value between 0 and .5 specifying the level of truncation}

\item{na.rm}{removes missing values of x}
}
\value{
The truncated geometric mean at (0.5, 0.95).
}
\description{
Computes the truncated geometric mean of a vector
}

